debugging capability
resolved all the issues

added the parent element in the submodules pom

fixed the issue of executing test multiple times

latest changes

edited the readme.md file with debugging capabilities setup steps

latest changes

added next feature details

Update README.md
