+++ execute/debug/testrunner.bat
@@ -3,6 +3,7 @@ setlocal enabledelayedexpansion
 set EXEC_ARGS=
 set SYSTEM_PARAMS=
 set NIMBUS_TAGS=
+cd ..\..
 :Loop
 IF "%~1"=="" GOTO Continue
     set arg=%~1
@@ -26,5 +27,6 @@ GOTO Loop
 RD /S /Q reports
 echo "mvn clean install"
 call mvn clean install -Pbuild
+@set MAVEN_DEBUG_OPTS=-Xdebug -Xrunjdwp:transport=dt_socket,server=y,suspend=y,address=9876
 echo "mvn exec:java -Pexecute -Dexec.args="-a -freports\req_res_files -tconfig\soap-ui\soapui-settings.xml %EXEC_ARGS% %SYSTEM_PARAMS%" %NIMBUS_TAGS%
 call mvn exec:java -Pexecute -Dexec.args="-a -freports\req_res_files -tconfig\soap-ui\soapui-settings.xml %EXEC_ARGS% %SYSTEM_PARAMS%" %NIMBUS_TAGS%
