+++ execute/debug/testrunner.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 SYSTEM_PARAMS=""
 EXEC_ARGS=""
-
+cd ..\..
 echo $JAVA_HOME
 
 for arg in "$@"
@@ -23,5 +23,6 @@ done
 rm -rf reports || true
 echo "mvn clean install"
 call mvn clean install -Pbuild
+export MAVEN_DEBUG_OPTS="-Xdebug -Xrunjdwp:transport=dt_socket,server=y,suspend=y,address=9876"
 echo "mvn exec:java -Pexecute -Dexec.args="-a -freports/req_res_files -tsoapui-settings.xml ${EXEC_ARGS}" ${SYSTEM_PARAMS}"
 mvn exec:java -Pexecute -Dexec.args="-a -freports/req_res_files -tconfig/soap-ui/soapui-settings.xml ${EXEC_ARGS}" ${SYSTEM_PARAMS}
